backend/app/cnn/preprocess.py
import cv2
import torch

def preprocess_image(path):
    img = cv2.imread(path, cv2.IMREAD_GRAYSCALE)
    if img is None:
        raise ValueError(f"Failed to load image. Check the path: {path}")

    img = cv2.resize(img, (224, 224))
    img = img / 255.0
    img = torch.tensor(img, dtype=torch.float32).unsqueeze(0).unsqueeze(0)
    return img
